pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "pdf-document-processor"
version = "1.0.0"
description = "Extract text/tables, merge/split PDFs, OCR, and generate professional reports"
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Your Name", email = "you@example.com" }]
license = { text = "MIT" }

dependencies = [
    "pypdf>=4.0.0",
    "pdfplumber>=0.10.0",
    "reportlab>=4.0.0",
    "python-docx>=1.1.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
ocr = [
    "pytesseract>=0.3.10",
    "pdf2image>=1.16.3",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=7.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "pylint>=3.0.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
pdf-processor = "main:main"

# ── Tool configs ──────────────────────────────────────────────────────────────

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "dist", "build"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pylint.messages_control]
disable = ["C0103", "R0903", "W0212"]

[tool.pylint.format]
max-line-length = 88

[tool.pytest.ini_options]
testpaths = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "smoke: quick end-to-end sanity checks",
    "slow: tests that take more than 1 second",
]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["pdf_processor", "report_generator"]
omit = ["test_*", "run_tests.py", "main.py"]

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 85
